# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Invoice & Receipt System API
  version: 0.1.0
  description: |
    Internal API for a small Israeli business operating as `עוסק פטור`.

    Conventions that hold everywhere:

    - Money is an **integer number of agorot** (1 ILS = 100 agorot). Never a
      decimal, never a float.
    - Timestamps are RFC 3339 in UTC. Business dates are `YYYY-MM-DD`, already
      interpreted in `Asia/Jerusalem`.
    - Authentication is a session cookie. Unsafe methods must also send the
      `X-CSRF-Token` header, echoing the `csrf_token` cookie.
    - Errors always use one envelope. `code` is a stable English identifier;
      `message` is Hebrew and is shown to the user as-is.

    Sending a document never affects the document: delivery runs in a
    background job, after issuance has committed, so a failed send cannot roll
    back an issued document (plan.md 14).

    Critical commands accept an `Idempotency-Key` header. A replay returns the
    original response with `Idempotent-Replay: true`, so a retry after a timeout
    never issues a second document with a second official number.
servers:
  - url: /api/v1
...
security:
  - sessionCookie: []
paths:
...
  /activities/:
    get:
      tags: [activities]
      summary: List activities, most recent first
      parameters:
        - $ref: '#/components/parameters/SearchQuery'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ActivityStatus' }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of activities
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActivityPage' }
    post:
      tags: [activities]
      summary: Create an activity
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActivityInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Activity' }
        '422': { $ref: '#/components/responses/Error' }
...
  /activities/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    get:
      tags: [activities]
      summary: One activity
      responses:
        '200':
          description: Activity
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Activity' }
        '404': { $ref: '#/components/responses/Error' }
    put:
      tags: [activities]
      summary: Replace an activity
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActivityInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Activity' }
        '422': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: |
        Set by `POST /auth/login`. Unsafe methods must also send
        `X-CSRF-Token`, echoing the `csrf_token` cookie.
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    SearchQuery:
      name: q
      in: query
      schema: { type: string }
      description: Case-insensitive substring match
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable English machine identifier
              examples: [UNAUTHORIZED, VALIDATION_FAILED, DOCUMENT_IMMUTABLE]
            message:
              type: string
              description: Hebrew text, shown to the user as-is
            details:
              type: object
              additionalProperties: true
              description: Field-level messages, keyed by field name
            request_id:
              type: string
...
    ActivityStatus:
      type: string
      enum: [PLANNED, ACTIVE, DONE, CANCELLED]
    ActivityInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        start_at: { type: [string, 'null'], format: date-time }
        end_at:
          type: [string, 'null']
          format: date-time
          description: Must not precede start_at; the database enforces this too.
        location: { type: string }
        status:
          allOf: [{ $ref: '#/components/schemas/ActivityStatus' }]
          default: PLANNED
        notes: { type: string }
    Activity:
      allOf:
        - $ref: '#/components/schemas/ActivityInput'
        - type: object
          properties:
            id: { type: string, format: uuid, readOnly: true }
            created_at: { type: string, format: date-time, readOnly: true }
            updated_at: { type: string, format: date-time, readOnly: true }
    ActivityPage:
      type: object
      properties:
        activities:
          type: array
          items: { $ref: '#/components/schemas/Activity' }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
...